refactor and remove middleware logic
